Refactored the election_provider_multiple_phase pallet
